% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageRdata.R
\docType{package}
\name{brainImageRdata}
\alias{brainImageRdata}
\alias{brainImageRdata-package}
\title{brainImageRdata}
\description{
brainImageRdata contains image masks for the developing human brain. It contains the data that supports the brainImageR library.
}
\section{data available}{

raw gene expression datasets from Allen Brain Atlas and image masks of the prenatal and adult brain. For details see brainImageRdata vignette.
}

